Fix setup scripts, update ruby, gems and js libs

+++ docker/docker-entrypoint.sh
@@ -2,8 +2,8 @@
 
 set -e
 
-if [ -f tmp/pids/server.pid ]; then
+if  [ "${1}" = "rails" ] && [ "${2}" = "server" ] && [ -f tmp/pids/server.pid ]; then
   rm tmp/pids/server.pid
 fi
 
-bundle exec rails s -b 0.0.0.0
+exec "${@}"
